refactor: tweaks
