+++ prices/europe.json
@@ -29,7 +29,7 @@
   "MT314": 3880,
   "MT315": 8990,
   "MT316": 8990,
-  "MT317": 11470,
+  "MT317": 10520,
   "MT320": 4990,
   "MT336": 1880,
   "MT337": 4990,
