+++ prices/usa.json
@@ -1,0 +1,30 @@
+{
+  "ACL1": 7000,
+  "APF2": 10000,
+  "CC02": 6500,
+  "CC04": 3500,
+  "FM3U": 2000,
+  "IN3BW": 1000,
+  "INB3C": 1500,
+  "MT301": 2490,
+  "MT302": 8900,
+  "MT303": 11490,
+  "MT305": 3900,
+  "MT308": 2490,
+  "MT314": 2490,
+  "MTS06": 20000,
+  "MTS08": 20000,
+  "MTX07": 10000,
+  "PBSB": 1000,
+  "PMBL": 2000,
+  "PMNG": 1000,
+  "PMSS": 2000,
+  "PPMR": 2000,
+  "PPSB": 1000,
+  "RFP2": 2000,
+  "W39B": 1500,
+  "W41B": 1500,
+  "WTSC": 2000,
+  "WTTC": 4500,
+  "WTUT": 5500
+}
