+++ scripts/prices/usa.js
@@ -1,0 +1,10 @@
+'use strict'
+
+const path = require('path')
+
+const { usa } = require('../../inventories')
+
+require('.')({
+  filepath: path.resolve(__dirname, '../../prices/usa.json'),
+  inventories: { usa }
+})
